+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add `tools` target for `Mage`.
 - Add application icons.
 - Generate Windows binary.
+- Add `stats` target for `Mage`. Compute project code statistics.
 
 ## [Unreleased] - 2026-02-06
 ### Added